+++ src/xyang/parser/statements/list.py
@@ -39,6 +39,7 @@ class ListStatementParser:
             kw.IF_FEATURE: self._parsers.parse_if_feature_stmt,
             kw.ANYDATA: self._parsers.parse_anydata,
             kw.ANYXML: self._parsers.parse_anyxml,
+            kw.NOTIFICATION: self._parsers.parse_notification,
             },
         )