+++ ts/src/parser/statements/index.ts
@@ -12,6 +12,7 @@ export { LeafStatementParser } from "./leaf";
 export { LeafListStatementParser } from "./leaf_list";
 export { ListStatementParser } from "./list";
 export { ModuleStatementParser } from "./module";
+export { NotificationStatementParser } from "./notification";
 export { MustStatementParser } from "./must";
 export { RefineStatementParser } from "./refine";
 export { RevisionStatementParser } from "./revision";